+++ package.json
@@ -12,6 +12,7 @@
     "react-dom": "18.2.0",
     "react-icons": "4.9.0",
     "react-redux": "8.0.7",
+    "react-router-dom": "6.13.0",
     "react-scripts": "2.1.3",
     "redux": "4.2.1",
     "redux-persist": "6.0.0",
