+++ src/App.css
@@ -1,9 +1,5 @@
 .App {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+ flex: 1;
 }
 
 #scrollTopButton {
