+++ src/pages/auth/Login.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export default function Login() {
+    console.log('login')
+    return (
+        <div><h1>Login</h1></div>
+    )
+}
